+++ apps/web/src/app/admin-console/organizations/members/services/member-actions/member-actions.types.ts
@@ -1,0 +1,13 @@
+import { OrganizationUserBulkResponse } from "@bitwarden/admin-console/common";
+
+export const REQUESTS_PER_BATCH = 500;
+
+export interface MemberActionResult {
+  success: boolean;
+  error?: string;
+}
+
+export class BulkActionResult {
+  successful: OrganizationUserBulkResponse[] = [];
+  failed: { id: string; error: string }[] = [];
+}
